Moved handling gists into new folder
